Curd opeartion added

+++ ServiceCenterWebApp.API/Models/Employee .cs
@@ -10,7 +10,7 @@ namespace ServiceCenterWebApp.API.Models
         public DateTime dob { get; set; }
         public int Experience { get; set; }
         public int Role { get; set; }
-        public int AdharID { get; set; }
+        public long AdharID { get; set; }
         public string Qualification { get; set; }
         public string Address { get; set; }
         public byte status { get; set; }
